+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.1.0]
+
+* Option to specify entity filter
+* Added option to specify filter for each feed exporter
+
 ## [2.0.0]
 
 * Extracted Google feed code into a separate package (`Geta.Optimizely.ProductFeed.Google`)